+++ changes.md
@@ -5,6 +5,9 @@
 * Added arch round
 * Added wave triangle
 * Added wave sawtooth
+* Fix arrow center
+  * 0 height returns line path
+* Chevron 0 height returns line path
 
 ## 0.0.8
 * Updated cadquery version